Create "articles" module, move all GC Articles code out of utils
Moving over functions related to GC Articles:
- find_item_url
- set_active_nav_item
- request_content
- get_nav_items

Also moved the tests over. Should function exactly as before.
