+++ tests/app/test_utils.py
@@ -12,8 +12,6 @@ from app.utils import (
     Spreadsheet,
     documentation_url,
     email_safe,
-    find_item_url,
-    set_active_nav_item,
     generate_next_dict,
     generate_notifications_csv,
     generate_previous_dict,
@@ -662,51 +660,3 @@ def test_get_template_with_html_allowed(mocker, app_, service_one, fake_uuid, al
 
     assert email_template is not None
     assert email_template.allow_html is allow_html
-
-
-def _get_items():
-    items = [{"name": "home", "url": "/"}, {"name": "page 1", "url": "/page-1"}, {"name": "page 2", "url": "/page-2"}]
-    return [i.copy() for i in items]
-
-
-def test_find_item_url():
-    items = _get_items()
-
-    found = find_item_url(items)
-    assert found is False
-
-    found = find_item_url(items, "")
-    assert found is False
-
-    found = find_item_url(items, "/")
-    assert found is True
-
-    found = find_item_url(items, "/page-2")
-    assert found is True
-
-    # trailing slash
-    found = find_item_url(items, "/page-2/")
-    assert found is False
-
-    found = find_item_url(items, "/page-2/page-1")
-    assert found is False
-
-    found = find_item_url(items, "/page-3")
-    assert found is False
-
-
-def test_set_active_nav_item():
-    items = _get_items()
-
-    set_active_nav_item(items, "/page-2")
-    matching_item = next((i for i in items if i["url"] == "/page-2"), None)
-    assert matching_item["active"] == True
-
-
-def test_set_no_active_nav_item():
-    items = _get_items()
-
-    # url doesn't exist in items
-    set_active_nav_item(items, "/page-3")
-    for item in items:
-        assert item["active"] == False
